fix: format-string bugs across ecm.c, pm1fs2.c, and others
Surfaced by adding __attribute__((format(printf,2,3))) to outputf and
auditing the warnings; the attribute itself is not part of this commit
(GMP doesn't annotate gmp_printf either, since the standard __printf__
archetype doesn't recognize %Zd, and there is no flag that suppresses
only the %Zd warnings while keeping the rest of -Wformat).

Real bugs fixed (the %Zd warnings -- about 110 of them -- are accepted
as a known limitation, same as in GMP itself):

  - ecm.c: %.0f used for uint64_t arguments (5 sites); cast to (double)
    to preserve the existing human-readable formatting.
  - pm1fs2.c: outputf(verbosity, prefix/suffix) called with non-literal
    format string (CWE-134-class issue under -Wformat-security); now
    outputf(verbosity, "%s", prefix/suffix).
  - pm1fs2.c: list_output_poly's "1\n"/"0\n" calls had spurious extra
    arguments; dropped them.
  - pm1fs2.c: choose_P trace had a "%zZd" typo for an mpz_t arg; fixed
    to "%Zd".
  - bestd.c: dF (unsigned long) was printed with %.0d; now %lu.
  - ecm2.c: ecm_rootsF index i (unsigned long) was printed with %d;
    now %lu.
  - pm1.c: mpz_sizeinbase result (size_t) was printed with %u; now %zu.
  - schoen_strass.c: lenA/lenB (unsigned int) were printed with %lu;
    now %u.
  - stage2.c: init_progression_coeffs printed unsigned-long d and e
    with %u, and unsigned-long d / mpz_gcd_ui result with %u; now %lu.

+++ ecm.c
@@ -1244,7 +1244,7 @@ ecm_stage1 (mpz_t f, mpres_t x, mpres_t A, mpmod_t n, double B1,
            if (mpres_is_zero (LCS_z[base_indx], n))
            {
              outputf (OUTPUT_VERBOSE, "Reached point at infinity, %.0f divides "
-                      "group orders\n", p);
+                      "group orders\n", (double) p);
              break;
            }
          }
@@ -1253,14 +1253,14 @@ ecm_stage1 (mpz_t f, mpres_t x, mpres_t A, mpmod_t n, double B1,
            if (mpres_is_zero (z, n))
            {
              outputf (OUTPUT_VERBOSE, "Reached point at infinity, %.0f divides "
-                      "group orders\n", p);
+                      "group orders\n", (double) p);
              break;
            }
          }
 
          if (stop_asap != NULL && (*stop_asap) ())
          {
-           outputf (OUTPUT_NORMAL, "Interrupted at prime %.0f\n", p);
+           outputf (OUTPUT_NORMAL, "Interrupted at prime %.0f\n", (double) p);
            break;
          }
 
@@ -1436,12 +1436,12 @@ ecm_stage1_W (mpz_t f, ell_curve_t E, ell_point_t P, mpmod_t n,
 	    }
 	    if (ell_point_is_zero (P, E, n)){
 		outputf (OUTPUT_VERBOSE, "Reached point at infinity, "
-			 "%.0f divides group orders\n", p);
+			 "%.0f divides group orders\n", (double) p);
 		break;
 	    }
-	    
+
 	    if (stop_asap != NULL && (*stop_asap) ()){
-		outputf (OUTPUT_NORMAL, "Interrupted at prime %.0f\n", p);
+		outputf (OUTPUT_NORMAL, "Interrupted at prime %.0f\n", (double) p);
 		break;
 	    }
